More controller stuff, hanving a tough time finding a place for everything.
Pushed the complicated loading step to the side for now, and got the
skeleton up. Thinking ui-nested-router is starting to look like the
mature step to take with the app's state transitions.

For now I'm not going to redirect the view state, since it's gonna be
really hard until we have something like ui-nested-router